Extract `Rodos/RealTime.hpp/ipp` into new `Time` component
