+++ Sts1CobcSw/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(Outcome)
 add_subdirectory(Periphery)
 add_subdirectory(Serial)
 add_subdirectory(Telemetry)
+add_subdirectory(Time)
 add_subdirectory(Utility)
 add_subdirectory(Vocabulary)
