+++ Sts1CobcSw/CobcSoftware/CMakeLists.txt
@@ -28,6 +28,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 Sts1CobcSw_Periphery
                 Sts1CobcSw_Serial
                 Sts1CobcSw_Telemetry
+                Sts1CobcSw_Time
                 Sts1CobcSw_Utility
                 Sts1CobcSw_Vocabulary
     )
