+++ Sts1CobcSw/Edu/CMakeLists.txt
@@ -9,6 +9,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
     target_sources(Sts1CobcSw_Edu PRIVATE Edu.cpp)
     target_link_libraries(
         Sts1CobcSw_Edu PRIVATE Sts1CobcSw_FileSystem Sts1CobcSw_Hal Sts1CobcSw_Periphery
+                               Sts1CobcSw_Time
     )
 endif()
 if(CMAKE_SYSTEM_NAME STREQUAL Linux)
