+++ Sts1CobcSw/FileSystem/LfsFlash.cpp
@@ -4,7 +4,7 @@
 #include <Sts1CobcSw/FileSystem/LfsMemoryDevice.hpp>  // IWYU pragma: associated
 #include <Sts1CobcSw/Periphery/Flash.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
-#include <Sts1CobcSw/Utility/RodosTime.hpp>
+#include <Sts1CobcSw/Time/RodosTime.hpp>
 
 #include <strong_type/difference.hpp>
