+++ Sts1CobcSw/Serial/CMakeLists.txt
@@ -1,1 +1,1 @@
-target_link_libraries(Sts1CobcSw_Serial INTERFACE Sts1CobcSw_Utility)
+target_link_libraries(Sts1CobcSw_Serial INTERFACE etl::etl)
