+++ Sts1CobcSw/Time/RealTime.hpp
@@ -12,4 +12,4 @@ namespace sts1cobcsw
 }
 
 
-#include <Sts1CobcSw/Utility/RealTime.ipp>  // IWYU pragma: keep
+#include <Sts1CobcSw/Time/RealTime.ipp>  // IWYU pragma: keep
