+++ iwyu.imp
@@ -58,9 +58,9 @@
   { include: ["\"Sts1CobcSw/Periphery/Fram.ipp\"",                      "private", "<Sts1CobcSw/Periphery/Fram.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Byte.ipp\"",                         "private", "<Sts1CobcSw/Serial/Byte.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Serial/Serial.ipp\"",                       "private", "<Sts1CobcSw/Serial/Serial.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Time/RealTime.ipp\"",                       "private", "<Sts1CobcSw/Time/RealTime.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Time/RodosTime.ipp\"",                      "private", "<Sts1CobcSw/Time/RodosTime.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/ErrorDetectionAndCorrection.ipp\"", "private", "<Sts1CobcSw/Utility/ErrorDetectionAndCorrection.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Utility/RealTime.ipp\"",                    "private", "<Sts1CobcSw/Utility/RealTime.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Utility/RodosTime.ipp\"",                   "private", "<Sts1CobcSw/Utility/RodosTime.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/Span.ipp\"",                        "private", "<Sts1CobcSw/Utility/Span.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/StringLiteral.ipp\"",               "private", "<Sts1CobcSw/Utility/StringLiteral.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Vocabulary/ProgramId.ipp\"",                "private", "<Sts1CobcSw/Vocabulary/ProgramId.hpp>", "public"] },
@@ -112,13 +112,13 @@
   { include: ["\"Sts1CobcSw/Serial/Serial.hpp\"",                                  "public", "<Sts1CobcSw/Serial/Serial.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Telemetry/TelemetryMemory.hpp\"",                      "public", "<Sts1CobcSw/Telemetry/TelemetryMemory.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Telemetry/TelemetryRecord.hpp\"",                      "public", "<Sts1CobcSw/Telemetry/TelemetryRecord.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Time/RealTime.hpp\"",                                  "public", "<Sts1CobcSw/Time/RealTime.hpp>", "public"] },
+  { include: ["\"Sts1CobcSw/Time/RodosTime.hpp\"",                                 "public", "<Sts1CobcSw/Time/RodosTime.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/Crc32Hardware.hpp\"",                          "public", "<Sts1CobcSw/Utility/Crc32Hardware.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/Crc32Software.hpp\"",                          "public", "<Sts1CobcSw/Utility/Crc32Software.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/DebugPrint.hpp\"",                             "public", "<Sts1CobcSw/Utility/DebugPrint.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/ErrorDetectionAndCorrection.hpp\"",            "public", "<Sts1CobcSw/Utility/ErrorDetectionAndCorrection.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/FlatArray.hpp\"",                              "public", "<Sts1CobcSw/Utility/FlatArray.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Utility/RealTime.hpp\"",                               "public", "<Sts1CobcSw/Utility/RealTime.hpp>", "public"] },
-  { include: ["\"Sts1CobcSw/Utility/RodosTime.hpp\"",                              "public", "<Sts1CobcSw/Utility/RodosTime.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/Span.hpp\"",                                   "public", "<Sts1CobcSw/Utility/Span.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Utility/StringLiteral.hpp\"",                          "public", "<Sts1CobcSw/Utility/StringLiteral.hpp>", "public"] },
   { include: ["\"Sts1CobcSw/Vocabulary/ProgramId.hpp\"",                           "public", "<Sts1CobcSw/Vocabulary/ProgramId.hpp>", "public"] },
